refactor: model과 핵심 로직을 담당하는 class 패키지로 구분
